📝 Add special scenario for Keymap-Drawer
## Changes Description

On a previous commit, I added a macro to the "toggle" key on the System
layer just to allow Keymap-Drawer to parse that key in a special way.

In this commit, I make that mechanism a bit better, and remove the
helper macro from the normal build, making Keymap-Drawer the only piece
that uses it.